Add exhaustive test suite and fix doc-tests
Test Suite:
- tests/common/mod.rs: Shared test utilities with API key handling
- tests/rest_integration.rs: 24 REST API integration tests covering
  aggregates, trades, quotes, snapshots, tickers, and pagination
- tests/ws_integration.rs: 22 WebSocket tests covering connection,
  authentication, subscriptions, and event streaming
- tests/mock_tests.rs: 30 mock-based unit tests for error handling
  and edge cases using wiremock

Model Fixes:
- Trade/Quote structs: Added serde aliases to support both v2 (short
  field names like 'p', 's') and v3 (long names like 'price', 'size')
  API response formats
- MinuteSnapshot: Changed accumulated_volume from u64 to f64

Doc-test Fixes:
- Converted all 'ignore' doc-tests to runnable examples
- src/parse.rs: Fixed JSON example with all required WsTrade fields
- src/rest/client.rs: Added complete stream example with imports
- src/rest/pagination.rs: Added complete PageStream example
- src/rest/request.rs: Made RestRequest and PaginatableRequest
  examples self-contained with inline type definitions

Test Results: 236 tests passing (0 ignored)
- 135 unit tests
- 30 mock tests
- 24 REST integration tests
- 22 WebSocket integration tests
- 25 doc-tests

+++ src/rest/endpoints/snapshots.rs
@@ -107,7 +107,7 @@ pub struct SnapshotTrade {
 pub struct MinuteSnapshot {
     /// Accumulated volume
     #[serde(rename = "av")]
-    pub accumulated_volume: Option<u64>,
+    pub accumulated_volume: Option<f64>,
     /// Open price
     #[serde(rename = "o")]
     pub open: f64,
